+++ src/app/categories/categories.html
@@ -16,4 +16,5 @@
     <a class="btn btn-info" ui-sref=".feedbacks">Feedbacks</a>
     <a class="btn btn-info" ui-sref=".events">Events</a>
 </div>
+
 <div ui-view></div>
